add: 캐러셀 인터벌타임 적용
