+++ src/pages/Main/carousel.scss
@@ -1,0 +1,76 @@
+.carousel {
+  display: flex;
+  align-items: center;
+
+  .prevImg {
+    width: 20px;
+    height: 20px;
+    background-color: white;
+    border: solid 2px black;
+    border-bottom: none;
+    border-right: none;
+    transform: rotate(-45deg);
+    margin-left: 3%;
+
+    &:hover {
+      cursor: pointer;
+    }
+  }
+  .nextImg {
+    width: 20px;
+    height: 20px;
+    background-color: white;
+    border: solid 2px black;
+    border-bottom: none;
+    border-left: none;
+    transform: rotate(45deg);
+    margin-right: 3%;
+
+    &:hover {
+      cursor: pointer;
+    }
+  }
+
+  .carouselWindow {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 10px auto 50px auto;
+    width: 1080px;
+    height: 650px;
+    overflow-x: hidden;
+    position: relative;
+
+    .images {
+      display: flex;
+      width: 100%;
+
+      .carouselImage {
+        width: 100%;
+        min-width: 1080px;
+        height: 585px;
+      }
+    }
+    .carouselBar {
+      display: flex;
+      justify-content: flex-start;
+      position: absolute;
+      bottom: 20%;
+
+      .barOn {
+        width: 5rem;
+        height: 7px;
+        background-color: white;
+        color: white;
+      }
+
+      .barOff {
+        width: 5rem;
+        height: 7px;
+        background-color: #e6e6e5;
+        opacity: 29%;
+        color: white;
+      }
+    }
+  }
+}
